Silence the unnecessary_new linter in third_party/dart/intl_translation's generated code.
The explicit "new"s cause lint warnings in generated code, which are errors for projects that opt-in to strict Dart lints.

This is the 2nd attempt at fixing this.
The first attempt in cl/211770033 was rolled back because it introduced a build breakage in the places in google3 that are still using Dart1.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=212265687