[dart][intl] Add message extraction check that enforces Intl.select() has the choice argument the first in args.
This is only mentioned in the contract [1], but not enforced in any way.
This causes problem when the order is different, and there is a translation, see [2], but not caught by unit test or even when using US locale on page (no translation) see [3].
This check is essential as this mistake is very hard to recognise.

[1] http://google3/third_party/dart/intl/lib/intl.dart?l=446
[2] https://screenshot.googleplex.com/7dsx87MAPye8SfP
[3] https://screenshot.googleplex.com/BJRJSB48ZzrHRy3

PiperOrigin-RevId: 427706066