Automated g4 rollback of changelist 427706066.
*** Reason for rollback ***

Breaks //ads/admob/appmgmt/client/app:generate_messages_extract_messages

https://fusion2.corp.google.com/presubmit/tap/427711379/OCL:427711379:BASE:427711412:1644495374329:1a3799a6;groups=PossiblyNewlyFailing/targets/invocations/4d36b506-fc30-4cb8-ae7a-839c07a29e3a/targets/%2F%2Fads%2Fadmob%2Fappmgmt%2Fclient%2Fapp:generate_messages_extract_messages/log

*** Original change description ***

[dart][intl] Add message extraction check that enforces Intl.select() has the choice argument the first in args.

This is only mentioned in the contract [1], but not enforced in any way.
This causes problem when the order is different, and there is a translation, see [2], but not caught by unit test or even when using US locale on page (no translation) see [3].
This check is essential as this mistake is very hard to recognise.

[1] http://google3/third_party/dart/intl/lib/intl.dart?l=446
[2] ht...

***

PiperOrigin-RevId: 427714973